listing videos implemented in to another component
